pyproject.toml
[project]
name = "nauphel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "crawl4ai>=0.6.3",
    "googlesearch-python>=1.3.0",
    "ipykernel>=6.29.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
